Update Cart Item functionality to customer

+++ app/models/cart_item.rb
@@ -10,16 +10,26 @@ class CartItem < ApplicationRecord
     menu_item.price * quantity
   end
 
+  def available?
+    menu_item.status
+  end
+
+  def self.available?
+    all.all? { |cart_item| cart_item.menu_item.status }
+  end
+
   def self.price
     all.map { |item| item.price }.sum()
   end
 
   def self.place_order(order)
-    all.each do |item|
-      OrderItem.create(order_id: order.id, name: item.name, quantity: item.quantity, price: item.price)
+    if all.available?
+      all.each do |item|
+        OrderItem.create(order_id: order.id, name: item.name, quantity: item.quantity, price: item.price)
+      end
+      order.price = all.price
+      order.save
+      all.destroy_all
     end
-    order.price = all.price
-    order.save
-    all.destroy_all
   end
 end
